+++ src/main/java/com/barlinc/unusual_prehistory/datagen/UP2BlockTagProvider.java
@@ -139,6 +139,16 @@ public class UP2BlockTagProvider extends BlockTagsProvider {
                 Blocks.GRASS_BLOCK
         );
 
+        this.tag(UP2BlockTags.JAWLESS_FISH_NIBBLING_BLOCKS).add(
+                MOSS_LAYER.get(),
+                MOSSY_LEPIDODENDRON_LOG.get(),
+                MOSSY_LEPIDODENDRON_WOOD.get(),
+                Blocks.SEAGRASS,
+                Blocks.TALL_SEAGRASS,
+                Blocks.MOSS_BLOCK,
+                Blocks.MOSS_CARPET
+        ).addTag(Tags.Blocks.COBBLESTONE_MOSSY);
+
         this.tag(UP2BlockTags.KENTROSAURUS_GRAZING_BLOCKS).add(
                 Blocks.GRASS_BLOCK
         );
@@ -147,6 +157,16 @@ public class UP2BlockTagProvider extends BlockTagsProvider {
                 Blocks.GRASS_BLOCK,
                 Blocks.MOSS_BLOCK
         );
+
+        this.tag(UP2BlockTags.TARTUOSTEUS_NIBBLING_BLOCKS).add(
+                MOSS_LAYER.get(),
+                MOSSY_LEPIDODENDRON_LOG.get(),
+                MOSSY_LEPIDODENDRON_WOOD.get(),
+                Blocks.SEAGRASS,
+                Blocks.TALL_SEAGRASS,
+                Blocks.MOSS_BLOCK,
+                Blocks.MOSS_CARPET
+        ).addTag(Tags.Blocks.COBBLESTONE_MOSSY);
 
         this.tag(UP2BlockTags.TELECREX_PECKING_BLOCKS).add(
                 Blocks.GRASS_BLOCK