+++ src/main/java/com/barlinc/unusual_prehistory/entity/Tartuosteus.java
@@ -1,5 +1,6 @@
 package com.barlinc.unusual_prehistory.entity;
 
+import com.barlinc.unusual_prehistory.entity.ai.goals.AquaticDigBlockGoal;
 import com.barlinc.unusual_prehistory.entity.ai.goals.CustomizableRandomSwimGoal;
 import com.barlinc.unusual_prehistory.entity.ai.goals.LargePanicGoal;
 import com.barlinc.unusual_prehistory.entity.ai.goals.TartuosteusGlideGoal;
@@ -7,6 +8,7 @@ import com.barlinc.unusual_prehistory.entity.base.PrehistoricAquaticMob;
 import com.barlinc.unusual_prehistory.entity.utils.LeapingMob;
 import com.barlinc.unusual_prehistory.registry.UP2Entities;
 import com.barlinc.unusual_prehistory.registry.UP2SoundEvents;
+import com.barlinc.unusual_prehistory.registry.tags.UP2BlockTags;
 import com.barlinc.unusual_prehistory.registry.tags.UP2EntityTags;
 import com.barlinc.unusual_prehistory.registry.tags.UP2ItemTags;
 import net.minecraft.core.BlockPos;
@@ -17,6 +19,7 @@ import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
+import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.FluidTags;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.DifficultyInstance;
@@ -59,11 +62,12 @@ public class Tartuosteus extends PrehistoricAquaticMob implements LeapingMob {
     @Override
     protected void registerGoals() {
         this.goalSelector.addGoal(0, new TryFindWaterGoal(this));
-        this.goalSelector.addGoal(1, new AvoidEntityGoal<>(this, LivingEntity.class, 6.0F, 2.0D, 2.0D, entity -> entity.getType().is(UP2EntityTags.TARTUOSTEUS_AVOIDS)));
-        this.goalSelector.addGoal(2, new TemptGoal(this, 1.2D, Ingredient.of(UP2ItemTags.TARTUOSTEUS_FOOD), false));
-        this.goalSelector.addGoal(3, new CustomizableRandomSwimGoal(this, 1.0D, 10));
-        this.goalSelector.addGoal(4, new TartuosteusGlideGoal(this));
-        this.goalSelector.addGoal(5, new LargePanicGoal(this, 2.0D, 10, 7));
+        this.goalSelector.addGoal(1, new LargePanicGoal(this, 2.0D, 10, 7));
+        this.goalSelector.addGoal(2, new AvoidEntityGoal<>(this, LivingEntity.class, 6.0F, 2.0D, 2.0D, entity -> entity.getType().is(UP2EntityTags.TARTUOSTEUS_AVOIDS)));
+        this.goalSelector.addGoal(3, new TemptGoal(this, 1.2D, Ingredient.of(UP2ItemTags.TARTUOSTEUS_FOOD), false));
+        this.goalSelector.addGoal(4, new AquaticDigBlockGoal(this, 30, 800, UP2BlockTags.TARTUOSTEUS_NIBBLING_BLOCKS, 1.0D));
+        this.goalSelector.addGoal(5, new CustomizableRandomSwimGoal(this, 1.0D, 10));
+        this.goalSelector.addGoal(6, new TartuosteusGlideGoal(this));
     }
 
     @Override