+++ src/main/java/com/barlinc/unusual_prehistory/entity/base/PrehistoricAquaticMob.java
@@ -12,10 +12,7 @@ import net.minecraft.tags.FluidTags;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.AnimationState;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.MobSpawnType;
-import net.minecraft.world.entity.MobType;
+import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.ai.navigation.PathNavigation;
 import net.minecraft.world.entity.animal.Bucketable;
 import net.minecraft.world.entity.player.Player;
@@ -26,6 +23,7 @@ import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.pathfinder.BlockPathTypes;
+import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.NotNull;
 
 @SuppressWarnings("deprecation")
@@ -115,7 +113,6 @@ public abstract class PrehistoricAquaticMob extends PrehistoricMob implements Bu
         if (!this.isInWaterOrBubble() && this.isAlive()) {
             if (this.onGround() && this.getRandom().nextFloat() < flopChance()) {
                 this.setDeltaMovement(this.getDeltaMovement().add((this.getRandom().nextFloat() * 2.0F - 1.0F) * 0.1F, 0.3D, (this.getRandom().nextFloat() * 2.0F - 1.0F) * 0.1F));
-//                this.setYRot(this.getRandom().nextFloat() * 360.0F);
                 this.playSound(this.getFlopSound(), this.getSoundVolume(), this.getVoicePitch());
             }
         }
@@ -173,6 +170,7 @@ public abstract class PrehistoricAquaticMob extends PrehistoricMob implements Bu
         compoundTag.putBoolean("Pacified", this.isPacified());
         compoundTag.putBoolean("FromEgg", this.isFromEgg());
         compoundTag.putLong("LastPoseTick", this.getLastPoseChangeTick());
+        compoundTag.putInt("EatingCooldown", this.getEatingCooldown());
     }
 
     @Override
@@ -185,6 +183,7 @@ public abstract class PrehistoricAquaticMob extends PrehistoricMob implements Bu
         this.setPacified(compoundTag.getBoolean("Pacified"));
         this.setFromEgg(compoundTag.getBoolean("FromEgg"));
         this.resetLastPoseChangeTick(compoundTag.getLong("LastPoseTick"));
+        this.setEatingCooldown(compoundTag.getInt("EatingCooldown"));
     }
 
     @Override